Use Tampere specific and city agnostic naming
Tamusers or even helusers is unnecessarily specific name for things
that can be common to any many organizations. Therefore this commit
uses city agnostic naming where possible. This is done in order to
minimize the need of renaming stuff from helusers to
tamusers everytime upstream changes are merged.

Similar changes can be suggested to upstream to avoid some of the
renaming altogether.

**Sidenote**

Whole package should be renamed to something like tunnistamo_users and
keep ecerything city agnostic. That is though much bigger job, because
it would most likely be a backwards incompatible change.