Handle duplicate or empty email addresses
When creating a new account, check if email already exists. We should
also avoid creating EmailAddress instance if the provider does not
include an email.

In cases where the payload is missing email, do not try to create an
EmailAddress instance. If another user already has this email address,
we should also not try and create a new EmailAddress instance.

In addition, if there is an existing SocialAccount matching another user,
raise an Authentication error instead of doing an assert.

Refs: TTVA-173